Added additional boundary tests

+++ Reviews/Controllers/ReviewsController.cs
@@ -117,6 +117,11 @@ namespace Reviews.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (id == null || id < 0)
+            {
+                return BadRequest();
+            }
+
             _repository.DeleteReview(id);
             await _repository.Save();
             return RedirectToAction(nameof(IndexAccount));
